+++ src/readabs/__init__.py
@@ -5,5 +5,5 @@ the Australian Bureau of Statistics (ABS)
 from . import readabs
 from .readabs import *
 
-__version__ = "0.0.12"
+__version__ = "0.0.13"
 __author__ = "Bryan Palmer"